requirements.txt
PyYaml
numpy
pandas
ultralytics
opencv-python
scikit-learn
tqdm